+++ Source/Analytics/Hardware/RAM/RAMInfo.cpp
@@ -106,7 +106,7 @@ size_t RAMInfo::GetAvailableRAM() const noexcept
 
 f32 RAMInfo::GetRAMUsage() noexcept
 {
-	f32 percentage = Map(
+	f32 percentage = MathUtils::Map(
 		static_cast<f32>(this->GetUsedRAM()), // value
 		0.0f, static_cast<f32>(this->GetAvailableRAM()), // from 0 to whatever ram available
 		0.0f, 100.0f // to 0 to 100 percentage range
@@ -116,7 +116,7 @@ f32 RAMInfo::GetRAMUsage() noexcept
 
 f32 RAMInfo::GetProcessRAMUsage() noexcept
 {
-	f32 percentage = Map(
+	f32 percentage = MathUtils::Map(
 		static_cast<f32>(this->GetProcessUsedRAM()), // value
 		0.0f, static_cast<f32>(this->GetAvailableRAM()), // from 0 to whatever ram available
 		0.0f, 100.0f // to 0 to 100 percentage range