+++ Source/Events/ApplicationEvent.hpp
@@ -5,6 +5,7 @@
 #include <Core/Core.hpp>
 #include <Core/Types.hpp>
 #include <sstream>
+#include <Utility/FileUtils.hpp> // fs::path
 
 #include "Event.hpp"
 
@@ -135,6 +136,23 @@ public:
 	EVENT_CLASS_CATEGORY(EventCategory::APPLICATION);
 };
 
+class WindowFileDropEvent : public Event
+{
+public:
+	explicit WindowFileDropEvent(const std::vector<fs::path>& filenames) 
+		:
+		m_filenames(filenames)
+	{}
+
+	constexpr const std::vector<fs::path>& GetFilenames() const noexcept { return m_filenames; }
+
+	EVENT_CLASS_TYPE(EventType::WINDOW_FILE_DROP);
+	EVENT_CLASS_CATEGORY(EventCategory::APPLICATION);
+
+private:
+	const std::vector<fs::path>& m_filenames;
+};
+
 class AppUpdateEvent : public Event
 {
 public:
@@ -161,7 +179,6 @@ public:
 	EVENT_CLASS_TYPE(EventType::APP_TICK);
 	EVENT_CLASS_CATEGORY(EventCategory::APPLICATION);
 };
-
 
 
 NS_ENIGMA_END